Layout changes in view review fragment

+++ app/src/main/java/com/navismart/navismart/model/ReviewModel.java
@@ -5,7 +5,7 @@ public class ReviewModel {
     private String review;
     private String reviewerName;
     private String reviewDate;
-    private int starRating;
+    private long starRating;
     private String bookingID;
     private String reviewerID;
     private long timeStamp;
@@ -56,11 +56,11 @@ public class ReviewModel {
         this.reviewDate = reviewDate;
     }
 
-    public int getStarRating() {
+    public long getStarRating() {
         return starRating;
     }
 
-    public void setStarRating(int starRating) {
+    public void setStarRating(long starRating) {
         this.starRating = starRating;
     }
